feat: 공통 hooks 폴더 구조 변경
