--- frontend/src/constants/modal.ts
@@ -1,4 +1,0 @@
-export const ADD_CATEGORY = 'add_category';
-export const DELETE_CATEGORY = 'delete_category';
-export const ADD_PRODUCT = 'add_product';
-export const DELETE_PRODUCT = 'delete_product';
