+++ frontend/src/hooks/index.ts
@@ -1,0 +1,4 @@
+export { default as useCategory } from './useCategory';
+export { default as useProduct } from './useProduct';
+export { default as useNumberInputs } from './useNumberInputs';
+export { default as useTextInptus } from './useTextInputs';
